Using publish-counts to record count of orders in order tabs
