+++ imports/plugins/core/orders/client/containers/ordersActionContainer.js
@@ -1,74 +1,11 @@
 import React from "react";
 import { Meteor } from "meteor/meteor";
+import { Counts } from "meteor/tmeasday:publish-counts";
 import { Orders } from "/lib/collections";
 import { Reaction, i18next } from "/client/api";
 import { composeWithTracker } from "/lib/api/compose";
 import OrderActions from "../components/orderActions";
 import * as Constants from "../../lib/constants";
-
-const OrderHelper =  {
-  makeQuery(filter) {
-    let query = {};
-
-    switch (filter) {
-      // New orders
-      case "new":
-        query = {
-          "workflow.status": "new"
-        };
-        break;
-
-      // Orders that have yet to be captured & shipped
-      case "processing":
-        query = {
-          "workflow.status": "coreOrderWorkflow/processing"
-        };
-        break;
-
-      // Orders that have been shipped, based on if the items have been shipped
-      case "shipped":
-        query = {
-          "items.workflow.status": "coreOrderItemWorkflow/shipped"
-        };
-        break;
-
-      // Orders that are complete, including all items with complete status
-      case "completed":
-        query = {
-          "workflow.status": "coreOrderWorkflow/completed",
-          "items.workflow.workflow": {
-            $in: ["coreOrderItemWorkflow/completed"]
-          }
-        };
-        break;
-
-      // Orders that have been captured, but not yet shipped
-      case "captured":
-        query = {
-          "billing.paymentMethod.status": "completed",
-          "shipping.shipped": false
-        };
-        break;
-
-      case "canceled":
-        query = {
-          "workflow.status": "canceled"
-        };
-        break;
-
-      // Orders that have been refunded partially or fully
-      case "refunded":
-        query = {
-          "billing.paymentMethod.status": "captured",
-          "shipping.shipped": true
-        };
-        break;
-      default:
-    }
-
-    return query;
-  }
-};
 
 function handleActionClick(filter) {
   Reaction.pushActionView({
@@ -94,7 +31,13 @@ function composer(props, onData) {
 
     filter.label = i18next.t(`order.filter.${filter.name}`, { defaultValue: filter.label });
     filter.i18nKeyLabel = `order.filter.${filter.name}`;
-    filter.count = Orders.find(OrderHelper.makeQuery(filter.name)).count();
+    if (filter.name === "new") {
+      filter.count = Counts.get("newOrder-count");
+    } else if (filter.name === "processing") {
+      filter.count = Counts.get("processingOrder-count");
+    } else if (filter.name === "completed") {
+      filter.count = Counts.get("completedOrder-count");
+    }
 
     return filter;
   });
