+++ server/publications/collections/orders.js
@@ -1,5 +1,78 @@
+import { Counts } from "meteor/tmeasday:publish-counts";
 import { Orders } from "/lib/collections";
 import { Reaction } from "/server/api";
+
+const OrderHelper =  {
+  makeQuery(filter) {
+    const shopId = Reaction.getShopId();
+    let query = {};
+
+    switch (filter) {
+      // New orders
+      case "new":
+        query = {
+          "shopId": shopId,
+          "workflow.status": "new"
+        };
+        break;
+
+      // Orders that have yet to be captured & shipped
+      case "processing":
+        query = {
+          "shopId": shopId,
+          "workflow.status": "coreOrderWorkflow/processing"
+        };
+        break;
+
+      // Orders that have been shipped, based on if the items have been shipped
+      case "shipped":
+        query = {
+          "shopId": shopId,
+          "items.workflow.status": "coreOrderItemWorkflow/shipped"
+        };
+        break;
+
+      // Orders that are complete, including all items with complete status
+      case "completed":
+        query = {
+          "shopId": shopId,
+          "workflow.status": "coreOrderWorkflow/completed",
+          "items.workflow.workflow": {
+            $in: ["coreOrderItemWorkflow/completed"]
+          }
+        };
+        break;
+
+      // Orders that have been captured, but not yet shipped
+      case "captured":
+        query = {
+          "shopId": shopId,
+          "billing.paymentMethod.status": "completed",
+          "shipping.shipped": false
+        };
+        break;
+
+      case "canceled":
+        query = {
+          "shopId": shopId,
+          "workflow.status": "canceled"
+        };
+        break;
+
+      // Orders that have been refunded partially or fully
+      case "refunded":
+        query = {
+          "shopId": shopId,
+          "billing.paymentMethod.status": "captured",
+          "shipping.shipped": true
+        };
+        break;
+      default:
+    }
+
+    return query;
+  }
+};
 
 /**
  * orders
@@ -14,6 +87,9 @@ Meteor.publish("Orders", function () {
     return this.ready();
   }
   if (Roles.userIsInRole(this.userId, ["admin", "owner"], shopId)) {
+    Counts.publish(this, "newOrder-count", Orders.find(OrderHelper.makeQuery("new")), { noReady: true });
+    Counts.publish(this, "processingOrder-count", Orders.find(OrderHelper.makeQuery("processing")), { noReady: true });
+    Counts.publish(this, "completedOrder-count", Orders.find(OrderHelper.makeQuery("completed")), { noReady: true });
     return Orders.find({
       shopId: shopId
     });
